refactor: extract AssetsTable
